\name{RetCalculatoR-package}
\alias{RetCalculatoR-package}
\alias{RetCalculatoR}
\docType{package}
\title{
\packageTitle{RetCalculatoR}
}
\description{
\packageDescription{RetCalculatoR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RetCalculatoR}
\packageIndices{RetCalculatoR}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{RetCalculatoR}

Maintainer: \packageMaintainer{RetCalculatoR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
